backend/requirements.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
rank_bm25
nltk
numpy